% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPriorToLog.R
\name{checkPriorToLog}
\alias{checkPriorToLog}
\title{Check if data values can be log transformed}
\usage{
checkPriorToLog(regData)
}
\arguments{
\item{regData}{data.frame to be checked for compatibility with log scale}
}
\description{
Check if all columns within a data.frame can be converted to logscale
}
\examples{
    checkPriorToLog(iris)
    testf <- data.frame(a = c(1, 2, 3, 0))
    checkPriorToLog(testDf)
 
}
