% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSetDummies.R
\name{checkSubstrings}
\alias{checkSubstrings}
\title{Helper function for makeSetDummies, assists in generating warning if any levels
within setValues are substrings or duplicates of other levels}
\usage{
checkSubstrings(strings)
}
\arguments{
\item{strings}{character vector, contains setValues argument from makeSetDummies}
}
\description{
Helper function for makeSetDummies, assists in generating warning if any levels
within setValues are substrings or duplicates of other levels
}
