% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFolders.R
\name{generateFolders}
\alias{generateFolders}
\title{Create 1 or more folders with identical subfolder structure}
\usage{
generateFolders(path, folders, subfolders)
}
\arguments{
\item{path}{character, location in which to create folders}

\item{folders}{character vector of length 1 or more.  Folders to be created in location path}

\item{subfolders}{character vector of lenth 0 or more.  Subfolders to be created within path/Folders.}
}
\description{
Generates one or more folders within a given path, all of which will contain
identical subfolders if specified.
}
\examples{
\dontrun{
# Create Folder structure with appropriate subfolders
generateFolders("C:/Users/KDV-Employee/Desktop/", 
               "P_testProj", c("Code", "Data"))
generateFolders("C:/Users/KDV-Employee/P_testProj/Data", 
               folders = c("Tidy", "Raw"), 
               subfolders = c("Docs", "Excel", "Figures", "Shapefle"))
}
}
