% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleEventsByVariable.R
\name{sampleEventsByVariable}
\alias{sampleEventsByVariable}
\title{Plot the frequency of sample data for different factor levels over time.}
\usage{
sampleEventsByVariable(df, setCol, varCol, var, dateBreaks, dateLimits = NULL)
}
\arguments{
\item{df}{Data frame or tibble in long format with a column with the variables
that will appear on y-axis, and a factored column containing options to
subset the data.}

\item{setCol}{Name of the column containing subset options as factor levels.}

\item{varCol}{Name of the column containing variables to plot on the y-axis.}

\item{var}{The character vector of names of variables to be used.
All variables should be present within the varCol column.}

\item{dateBreaks}{The date breaks that will be used on the x-axis.}

\item{dateLimits}{The date limits on the y-axis.}
}
\value{
NULL, plots are printed for each factor level of setCol.
}
\description{
This function takes a data frame and creates a plot with time on the x-axis
and variables on the y-axis to show the spread of sampling events over time.
This function also takes a column containing subset options as different
factor levels to generate a new plot for each subset.
}
\examples{
\dontrun{
 airqualLong <- tidyr::pivot_longer(airquality, 1:4,names_to = "vars", values_to = "val")
 airqualLong$DATE <- paste(airqualLong$Month, airqualLong$Day, "2010", sep = "/")
 airqualLong$DATE <- as.Date(airqualLong$DATE, tryFormats = c("\%m/\%d/\%Y"))
 airqualLong$Month <- as.factor(airqualLong$Month)
 
 sampleEventsByVariable(df = airqualLong, setCol = "Month", varCol = "vars", var = unique(airqualLong$vars), dateBreaks = "1 day")
 }
}
