% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPhrase.R
\name{setPhrase}
\alias{setPhrase}
\title{Turns a vector of values into a sentence style list,
inserting "and" prior to the last item, keeps each item separate within vector}
\usage{
setPhrase(setVector, oxford = TRUE, vectorize = FALSE)
}
\arguments{
\item{setVector}{character vector, vector of values to be converted to sentence style}

\item{oxford}{logical, if TRUE adds an oxford comma when applicable}

\item{vectorize}{logical, indicates whether to split phrase items into seperate
indices within output vector.  This may be useful if output is being used within HTML bulleted list.}
}
\value{
a character vector containing input items formatted as a sentance
}
\description{
Turns a vector of values into a sentence style list,
inserting "and" prior to the last item, keeps each item separate within vector
}
