% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitMarkdownTutorial.R
\name{splitMarkdownTutorial}
\alias{splitMarkdownTutorial}
\title{Generate multiple R files from one R markdown file}
\usage{
splitMarkdownTutorial(file = "Walkthrough_01_CombinedExercises.Rmd")
}
\arguments{
\item{file}{The path to the RMD file you'd like to split up.}
}
\value{
Multiple R files from each section of the R markdown document
}
\description{
This function takes a single RMD document containing headers (headers MUST be
denoted by "# ", for example "# CORE ELEMENTS) and breaks all sections into
separate R files in your working directory. In these R files, text outside of
code chunks appear as Roxygen "#' " comments, and chunks of code are denoted
"## ----chunkName--".
}
\examples{
  file <- "Walkthrough_01_CombinedExercises.Rmd"
  splitMarkdownTutorial(file)

}
