% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSetDummies.R
\name{makeSetDummies}
\alias{makeSetDummies}
\title{Generates a data.frame of dummy columns from a data.frame containing specified
columns and values.  Often input is from from Arc Collector when
user entered values from a multi-select widget}
\usage{
makeSetDummies(df, setColumns, setValues, numeric = FALSE)
}
\arguments{
\item{df}{data.frame containing column(s) from which to generate dummy columns}

\item{setColumns}{character vector specifying which columns in df contain values from which to generate dummy columns}

\item{setValues}{character vector specifying which values in setColumns to use to generate dummy columns}

\item{numeric}{logical, if FALSE dummy columns will be of class logical, if TRUE dummy columns will be numeric}
}
\value{
data.frame
}
\description{
Generates a data.frame of dummy columns from a data.frame containing specified
columns and values.  Often input is from from Arc Collector when
user entered values from a multi-select widget
}
