% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPhrase.R
\name{setPhrase}
\alias{setPhrase}
\title{Author: Ashton Drew
Turns a vector of values into a sentence style list,
inserting "and" prior to the last item.}
\usage{
setPhrase(setVector, oxford = TRUE)
}
\arguments{
\item{setVector}{the vector of values to be converted to sentence style}

\item{oxford}{do you want an oxford comma? default is TRUE}
}
\value{
a single character string
}
\description{
Author: Ashton Drew
Turns a vector of values into a sentence style list,
inserting "and" prior to the last item.
}
